trading_concierge_mo/tradeQueryApi/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-multipart==0.0.22
requests>=2.32.5
pytest==8.0.0
python-dateutil